Refactor Layout subcomponents.
- Create Hero component and dynamically import in Header
- Extract MobileMenu and Search buttons to MobileMenu component and dynamically import in Header
- Dynamically import MainMenu in Header
- Refactor UserNav and subcomponents and dynamically import in Header
- Dynamically import GlobalStyles in Layout component
- Extract styled component from Alerts component
