+++ components/Layout/Header/Header.tsx
@@ -1,146 +1,41 @@
-import { useState } from "react"
+import dynamic from "next/dynamic"
 import Headroom from "react-headroom"
-import { MenuIcon, SearchIcon } from "@heroicons/react/outline"
 
-import MobileMenu from "./MobileMenu"
-import TopNav from "./TopNav"
-import MainMenu from "./MainMenu"
-import UserNav from "./UserNav"
 import { ArtilleryPage_Acfhome_Hero, Maybe } from "@api/gql/types"
 
-const navigation = {
-  categories: [
-    {
-      name: "Women",
-      featured: [
-        {
-          name: "New Arrivals",
-          href: "#",
-          imageSrc:
-            "https://tailwindui.com/img/ecommerce-images/mega-menu-category-01.jpg",
-          imageAlt:
-            "Models sitting back to back, wearing Basic Tee in black and bone.",
-        },
-        {
-          name: "Basic Tees",
-          href: "#",
-          imageSrc:
-            "https://tailwindui.com/img/ecommerce-images/mega-menu-category-02.jpg",
-          imageAlt:
-            "Close up of Basic Tee fall bundle with off-white, ochre, olive, and black tees.",
-        },
-        {
-          name: "Accessories",
-          href: "#",
-          imageSrc:
-            "https://tailwindui.com/img/ecommerce-images/mega-menu-category-03.jpg",
-          imageAlt:
-            "Model wearing minimalist watch with black wristband and white watch face.",
-        },
-        {
-          name: "Carry",
-          href: "#",
-          imageSrc:
-            "https://tailwindui.com/img/ecommerce-images/mega-menu-category-04.jpg",
-          imageAlt:
-            "Model opening tan leather long wallet with credit card pockets and cash pouch.",
-        },
-      ],
-    },
-    {
-      name: "Men",
-      featured: [
-        {
-          name: "New Arrivals",
-          href: "#",
-          imageSrc:
-            "https://tailwindui.com/img/ecommerce-images/mega-menu-01-men-category-01.jpg",
-          imageAlt:
-            "Hats and sweaters on wood shelves next to various colors of t-shirts on hangers.",
-        },
-        {
-          name: "Basic Tees",
-          href: "#",
-          imageSrc:
-            "https://tailwindui.com/img/ecommerce-images/mega-menu-01-men-category-02.jpg",
-          imageAlt: "Model wearing light heather gray t-shirt.",
-        },
-        {
-          name: "Accessories",
-          href: "#",
-          imageSrc:
-            "https://tailwindui.com/img/ecommerce-images/mega-menu-01-men-category-03.jpg",
-          imageAlt:
-            "Grey 6-panel baseball hat with black brim, black mountain graphic on front, and light heather gray body.",
-        },
-        {
-          name: "Carry",
-          href: "#",
-          imageSrc:
-            "https://tailwindui.com/img/ecommerce-images/mega-menu-01-men-category-04.jpg",
-          imageAlt:
-            "Model putting folded cash into slim card holder olive leather wallet with hand stitching.",
-        },
-      ],
-    },
-  ],
-  pages: [
-    { name: "Company", href: "#" },
-    { name: "Stores", href: "#" },
-  ],
-}
+// import TopNav from "./TopNav"
 
-function classNames(...classes: string[]) {
-  return classes.filter(Boolean).join(" ")
-}
+// ####
+// #### Dynamic Imports
+// ####
+
+const importOpts = {}
+
+const Hero = dynamic(() => import("./Hero"), {})
+const MainMenu = dynamic(() => import("./MainMenu"), {})
+const MobileMenu = dynamic(() => import("./MobileMenu"), {})
+const UserNav = dynamic(() => import("./UserNav"), {
+  ssr: false,
+})
+
+// ####
+// #### Types
+// ####
 
 type PropsType = {
   hero?: Maybe<ArtilleryPage_Acfhome_Hero> | undefined
 }
 
+// ####
+// #### Component
+// ####
+
 export const Header = ({ hero }: PropsType) => {
-  const [mobileMenuOpen, setMobileMenuOpen] = useState<boolean>(false)
-  const [signInOpen, setSignInOpen] = useState<boolean>(false)
-  const [searchOpen, setSearchOpen] = useState<boolean>(false)
+  // const [searchOpen, setSearchOpen] = useState<boolean>(false)
 
   return (
     <>
-      <MobileMenu
-        open={mobileMenuOpen}
-        setOpen={setMobileMenuOpen}
-        navigation={navigation}
-        classNames={classNames}
-      />
-
       <div className="relative bg-gray-900">
-        {/* Decorative image and overlay */}
-        {hero && (
-          <>
-            {hero.image?.sourceUrl && (
-              <div
-                aria-hidden="true"
-                className="absolute inset-0 overflow-hidden"
-              >
-                <img
-                  src={hero.image.sourceUrl}
-                  alt=""
-                  className="w-full h-full object-center object-cover"
-                />
-              </div>
-            )}
-            {hero.overlay && (
-              <div
-                aria-hidden="true"
-                className="absolute inset-0"
-                style={{
-                  backgroundColor: hero.overlay.color || "#111827",
-                  opacity: hero.overlay.opacity || 0.5,
-                }}
-              />
-            )}
-          </>
-        )}
-
         {/* Navigation */}
         <header className="relative z-10">
           <Headroom
@@ -168,28 +63,10 @@ export const Header = ({ hero }: PropsType) => {
                         </a>
                       </div>
 
-                      <MainMenu
-                        navigation={navigation}
-                        classNames={classNames}
-                      />
+                      <MainMenu />
 
                       {/* Mobile menu and search (lg-) */}
-                      <div className="flex-1 flex items-center lg:hidden">
-                        <button
-                          type="button"
-                          className="-ml-2 p-2 text-white"
-                          onClick={() => setMobileMenuOpen(true)}
-                        >
-                          <span className="sr-only">Open menu</span>
-                          <MenuIcon className="h-6 w-6" aria-hidden="true" />
-                        </button>
-
-                        {/* Search */}
-                        <a href="#" className="ml-2 p-2 text-white">
-                          <span className="sr-only">Search</span>
-                          <SearchIcon className="w-6 h-6" aria-hidden="true" />
-                        </a>
-                      </div>
+                      <MobileMenu />
 
                       {/* Logo (lg-) */}
                       <a href="#" className="lg:hidden">
@@ -201,10 +78,7 @@ export const Header = ({ hero }: PropsType) => {
                         />
                       </a>
 
-                      <UserNav
-                        iconSize="h-6 w-6"
-                        setSignInOpen={setSignInOpen}
-                      />
+                      <UserNav />
                     </div>
                   </div>
                 </div>
@@ -213,26 +87,9 @@ export const Header = ({ hero }: PropsType) => {
           </Headroom>
         </header>
 
-        {hero && (
-          <div className="relative max-w-3xl mx-auto py-32 px-6 flex flex-col items-center text-center sm:py-64 lg:px-0">
-            {hero.title && (
-              <h1 className="text-4xl font-extrabold tracking-tight text-white lg:text-6xl">
-                {hero.title}
-              </h1>
-            )}
-            {hero.text && (
-              <p className="mt-4 text-xl text-white">{hero.text}</p>
-            )}
-            {hero.link && (
-              <a
-                href={hero.link.url || "#"}
-                className="mt-8 inline-block bg-white border border-transparent rounded-md py-3 px-8 text-base font-medium text-gray-900 hover:bg-gray-100"
-              >
-                {hero.link.label}
-              </a>
-            )}
-          </div>
-        )}
+        {/* Decorative image and overlay */}
+
+        {hero && <Hero hero={hero} />}
       </div>
     </>
   )
